fix feedUDPPortNum type
> variable(feedUDPPortNum, int)

expects C 'int'.  If sizeof() is too small,
then some other global may be clobbered.

+++ src/driver/device.h
@@ -272,6 +272,6 @@ struct epicsShareClass Device : public epicsThreadRunable
 extern int feedNumInFlight;
 extern double feedTimeout;
 extern int feedUDPHeaderSize;
-extern unsigned short feedUDPPortNum;
+extern int feedUDPPortNum;
 
 #endif // DEVICE_H
